Algorithm.java: Checking the dictionary conflicts
Check whether we have conflicts in the dictionary or not.
for every stroke we should count how many times it appears, if it appears only once it's good. 
But if more than once it means that for some Arabic characters we may have more than one stroke and it is not good.
So, we are displaying here the stroke and the number for appearance in the outline dictionary.